Salvando para no perder el progreso Se agregaron misiles seguidores en torretas grandes y el jugador Se agregaron los niveles 4 a los disparos laser multiplicador y wave Se corrigieron algunos bugs graficos Se incremento la vida a las naves Se implemento la colision entre disparos de enemigo y jugador Se aumento la velocidad del misil bomba Se aumento la velocidad de spam y la velocidad de la accion disparar de los enemigos Se corrigieron varios bugs del laser se agrego un nuevo compañero como evolucion del primero (ahora si agarras 2 veces el compañero tendras 2 compañeros) Se cambio el sonido al agarrar un escudo y se agrego sonido de alerta cuando el escudo esta por debajo de 50 de hp Faltan corregir algunas cosas aun

+++ src/ProyectoX/Disparos/Wave/DisparoWave.java
@@ -17,7 +17,7 @@ public class DisparoWave extends DisparoJugador {
 
 	private static final URL url = Disparo.class.getClassLoader().getResource("ProyectoX/img/Disparos/Wave/wave.png");
 		
-	protected int variacion = 0;
+	protected double variacion = 0;
 	
 	/**
 	 * Constructor de la clase DisparoWave
